Feat : 검색기능 오류 수정 완료 Refactor : Detail page, Qna Board code Refactor Design : footer 고정 Fix : pagenation 데이터 없을 시 버튼 안 보이게 수정완료, 데이터 검색 하면 전체 페이지 수만큼 나온 현상 수정
